feat(pg): make PostgreSQL opt-in via `require "tep/pg"` (#216)
PG reached every compiled app's closure through method calls in core
Broadcast/Presence + the lib/tep.rb seed block, so a non-PG app linked
libpq and the 74 tep_pg_* FFI symbols it never used -- and, on current
master, miscompiled on pg.rb's `opts.each` poly-receiver (the sole
remaining blocker to a building `toy serve`, per spinel-dev).

Make PG opt-in:
- lib/tep.rb no longer require_relative's tep/pg; the PG seed clusters
  (broadcast_pg/presence_pg setters, enable_pg_backend/poll_pg_once,
  the mirror seeds, and the PG::Connection/Result/Pool type-seeding)
  move to lib/tep/pg.rb's load path.
- Core Broadcast.publish calls a no-op `cross_worker_notify`; core
  Presence track/untrack/set_status call no-op mirror_insert/delete/
  status stubs. lib/tep/pg.rb REDEFINES these (last-definition-wins,
  verified at f6d5eef + master) with the real LISTEN/NOTIFY + mirror
  exec_params, alongside the relocated PG cmeths.
- App keeps the PG ivars + setters; in a non-PG app the setters are
  uncalled and DCE away.
- bin/tep splices lib/tep/pg.rb in on `require "tep/pg"` (and runs the
  @TEP_PG_*@ placeholder subst the wholesale-core path does); a non-PG
  app never names it. PG tests/examples add the require.

Validated:
- non-PG hello/sinatra_style: built at pin (f6d5eef) AND master
  (5959b64) with 0 tep_pg_ symbols + no libpq linkage (DCE win).
- PG pg_hello: built at pin with 63 tep_pg_ symbols (opt-in works).
- test_broadcast 15/15, test_presence 24/24, test_live_view 17/17;
  PG tests skip cleanly without PG_TEST_URL.

Unblocks toy serve on fixed-master without waiting on the pg.rb
opts.each upstream fix (which is now moot for non-PG consumers).
