+++ test/test_broadcast_pg.rb
@@ -18,6 +18,7 @@ class TestBroadcastPg < TepTest
 
   app_source <<~RB
     require 'sinatra'
+    require "tep/pg"          # opt-in PG backend (#216)
 
     PG_URL  = "#{PG_URL}"
     CHANNEL = "#{CHANNEL}"
